Move out transactions related structs into separate module
