+++ core/shard/Cargo.toml
@@ -12,5 +12,6 @@ serde_derive = "1.0"
 parking_lot = "0.7.1"
 
 primitives = { path = "../primitives" }
+transaction = { path = "../transaction" }
 storage = { path = "../storage" }
 chain = { path = "../chain" }
