+++ core/transaction/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "transaction"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[dependencies]
+byteorder = "1.2"
+serde = "1.0"
+serde_derive = "1.0"
+
+primitives = { path = "../primitives" }
