+++ node/cli/Cargo.toml
@@ -20,6 +20,7 @@ node-runtime = { path = "../runtime" }
 node-http = { path = "../http" }
 parking_lot = "0.7.1"
 primitives = { path = "../../core/primitives" }
+transaction = { path = "../../core/transaction" }
 storage = { path = "../../core/storage" }
 txflow = { path = "../../core/txflow" }
 client = { path = "../../node/client" }
