+++ node/consensus/Cargo.toml
@@ -9,6 +9,7 @@ network = { path = "../network" }
 beacon = { path = "../../core/beacon" }
 client = { path = "../../node/client" }
 primitives = { path = "../../core/primitives" }
+transaction = { path = "../../core/transaction" }
 substrate-network-libp2p = { git = "https://github.com/nearprotocol/substrate", rev = "2db35a0c6e87fc4450a77da548ef6e2db8b0d057" }
 beacon-chain-handler = { path = "../beacon-chain-handler" }
 txflow = { path = "../../core/txflow" }
