+++ node/consensus/src/passthrough.rs
@@ -2,7 +2,8 @@ use client::ChainConsensusBlockBody;
 use futures::sync::mpsc::{Receiver, Sender};
 use futures::{future, Future, Sink, Stream};
 use primitives::signature::DEFAULT_SIGNATURE;
-use primitives::types::{MessageDataBody, SignedMessageData, ChainPayload, Gossip};
+use primitives::types::{MessageDataBody, SignedMessageData, Gossip};
+use transaction::ChainPayload;
 use std::collections::HashSet;
 use tokio;
 use txflow::txflow_task::beacon_witness_selector::BeaconWitnessSelector;
