+++ node/http/Cargo.toml
@@ -14,4 +14,5 @@ serde_json = "1.0"
 beacon = { path = "../../core/beacon" }
 node-runtime = { path = "../runtime" }
 primitives = { path = "../../core/primitives" }
+transaction = { path = "../../core/transaction" }
 shard = { path = "../../core/shard" }
