+++ node/network/Cargo.toml
@@ -22,6 +22,7 @@ beacon = { path = "../../core/beacon" }
 libp2p = { git = "https://github.com/tomaka/libp2p-rs", rev = "997d0163bc8a7e11559524ad8466bc3b1850c8ec", default-features = false, features = ["secio-rsa", "secio-secp256k1"] }
 substrate-network-libp2p = { git = "https://github.com/nearprotocol/substrate", rev = "2db35a0c6e87fc4450a77da548ef6e2db8b0d057" }
 primitives = { path = "../../core/primitives" }
+transaction = { path = "../../core/transaction" }
 chain = { path = "../../core/chain" }
 storage = { path = "../../core/storage", features=["test-utils"] }
